feat: Card CRUD 구현
Card entity CRUD 구현, 링크드리스트 없음

+++ src/main/java/org/sparta/hanghae99trello/dto/CommentResponseDto.java
@@ -8,7 +8,7 @@ import org.sparta.hanghae99trello.entity.Comment;
 @Setter
 public class CommentResponseDto {
 
-    private int commentId;
+    private Long commentId;
     private String commentMessage;
 
     public CommentResponseDto(Comment comment) {
